Backend/requirements.v2.txt
asyncio==3.4.3
bcrypt==4.0.1
black==23.9.1
cachetools==5.3.2
certifi==2023.7.22
cryptography==41.0.4
dateparser==1.1.8
deepdiff==6.5.0
deepmerge==1.1.0
fastapi==0.99.1
fastapi-jwt-auth==0.5.0
FastAPI-SQLAlchemy==0.2.1
flatdict==4.0.1
GeoAlchemy2==0.14.2
isort==5.12.0
loguru==0.7.2
marshmallow==3.20.1
marshmallow-mongoengine==0.31.2
marshmallow-sqlalchemy==0.29.0
mongoengine==0.27.0
nested-lookup==0.2.25
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
pdfkit==1.0.0
pylint==2.17.6
PyPDF2==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
py3-wkhtmltopdf==0.4.1
reportlab==4.0.9
requests==2.31.0
shapely==2.0.2
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
uvicorn==0.23.2
xmltodict==0.13.0